abstract some filtering logic into a composable
